Relocate billing component files
